+++ Libraries/Recipe.cs
@@ -2,16 +2,16 @@
 {
     public struct Recipe
     {
-        public int Level { get; set; }
-        public int RLevel { get; init; }
+        public byte Level { get; set; }
+        public short RLevel { get; init; }
         public int Difficulty { get; init; }
-        public int Durability { get; init; }
+        public byte Durability { get; init; }
         public int StartQuality { get; init; }
         public int MaxQuality { get; init; }
         public bool IsExpert { get; set; }
 
-        public int ProgressDivider { get; init; }
-        public int QualityDivider { get; init; }
+        public byte ProgressDivider { get; init; }
+        public byte QualityDivider { get; init; }
         public double ProgressModifier { get; init; }
         public double QualityModifier { get; init; }
     }
